endpoints file moved
